use only placeholder files
